% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroconpg.R
\name{create_heroku_pgurl_args}
\alias{create_heroku_pgurl_args}
\title{Create arguments for the heroku CLI}
\usage{
create_heroku_pgurl_args(prop, app)
}
\arguments{
\item{prop}{Property to look for}

\item{app}{Target heroku app}
}
\description{
Create arguments for the heroku CLI
}
