% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroconpg.R
\name{get_con}
\alias{get_con}
\title{Get con object to Heroku postgres db from heroku parameters}
\usage{
get_con(prop, app, os = "unix")
}
\arguments{
\item{prop}{Property to look for}

\item{app}{Target Heroku app}

\item{os}{Target Operating System}
}
\description{
Get con object to Heroku postgres db from heroku parameters
}
