% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroconpg.R
\name{get_con_from_url}
\alias{get_con_from_url}
\title{Get con object to Heroku postgres db from url}
\usage{
get_con_from_url(db_url)
}
\arguments{
\item{db_url}{URL of Postgres DB}
}
\description{
Get con object to Heroku postgres db from url
}
