% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroconpg.R
\name{get_heroku_pg_url}
\alias{get_heroku_pg_url}
\title{Get postgres db url from Heroku}
\usage{
get_heroku_pg_url(prop, app, os = "unix")
}
\arguments{
\item{prop}{Property to look for}

\item{app}{Target Heroku app}

\item{os}{Target Operating System}
}
\description{
Get postgres db url from Heroku
}
