% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroconpg.R
\name{run_linux_cmd}
\alias{run_linux_cmd}
\title{Run command on LINUX os}
\usage{
run_linux_cmd(heroku_cli = "heroku", arguments)
}
\arguments{
\item{heroku_cli}{Path of the Heroku CLI}

\item{arguments}{Vector of arguments for CLI}
}
\description{
Run command on LINUX os
}
