% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroconpg.R
\name{run_unix_cmd}
\alias{run_unix_cmd}
\title{Run command on UNIX os}
\usage{
run_unix_cmd(heroku_cli = "/snap/bin/heroku", arguments)
}
\arguments{
\item{heroku_cli}{Path of the Heroku CLI}

\item{arguments}{Vector of arguments for CLI}
}
\description{
Run command on UNIX os
}
